chat-app/whatsapp-selenium/requirements.txt
selenium>=4.15.0
motor>=3.3.0
pymongo>=4.6.0
Pillow>=10.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
